+++ packages/console-types/src/types/project.test.ts
@@ -234,6 +234,9 @@ t.test('project validated', t => {
         viewBasicHomepage: true,
       },
       imagePullSecretNames: ['some-imagePullSecret-here', 'some-other-one'],
+      containerRegistriesSettings: {
+        disabledProviderIds: ['some', 'disabled', 'providers'],
+      },
     }
 
     t.ok(validate(project), validationMessage(validate.errors))
@@ -362,6 +365,9 @@ t.test('project validated', t => {
         viewBasicHomepage: true,
       },
       imagePullSecretNames: ['some-imagePullSecret-here', 'some-other-one'],
+      containerRegistriesSettings: {
+        disabledProviderIds: ['some', 'disabled', 'providers'],
+      },
     }
 
     t.ok(validate(projectObj), validationMessage(validate.errors))