+++ packages/console-types/src/types/project.ts
@@ -16,6 +16,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+// TODO: create subfolder project/ and split entities in multiple files?
+/* eslint-disable max-lines*/
+
 import { FromSchema } from 'json-schema-to-ts'
 
 import { DASHBOARD_TYPES, DASHBOARD_TYPE_IFRAME } from '../constants/dashboard'
@@ -521,6 +524,17 @@ export const project = {
     },
     monitoring,
     configurationManagement,
+    containerRegistriesSettings: {
+      type: 'object',
+      properties: {
+        disabledProviderIds: {
+          type: 'array',
+          items: {
+            type: 'string',
+          },
+        },
+      },
+    },
   },
   required: [
     '_id',